Move parse_headers to separate module
Since the parse_headers is being unittested, but is implemented in the
start.py module, that set of unittests have a nasty side effect, somehow
always making the test process returning a 143 code (sigterm).
By extracting it, this side effect is removed from the test module.
