+++ tests/usecase/test_custom_headers.py
@@ -1,6 +1,6 @@
 import unittest
 import os
-import start
+import headers
 import json
 
 
@@ -10,7 +10,7 @@ class TestCaseCustomHeaderConfig(unittest.TestCase):
             {"X-Frame-Options": "allow-from https://mendix.com"}
         )
         os.environ["X_FRAME_OPTIONS"] = "deny"
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertIn(
             "add_header X-Frame-Options 'allow-from https://mendix.com';",
             header_config,
@@ -20,20 +20,20 @@ class TestCaseCustomHeaderConfig(unittest.TestCase):
         os.environ["HTTP_RESPONSE_HEADERS"] = json.dumps(
             {"X-Frame-Options": "allow-form htps://mendix.com"}
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertEquals("", header_config)
 
     def test_valid_with_xframeOption(self):
         os.environ["HTTP_RESPONSE_HEADERS"] = "{}"
         os.environ["X_FRAME_OPTIONS"] = "DENY"
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertIn("add_header X-Frame-Options 'DENY';", header_config)
 
     def test_valid_header_referrerPolicy(self):
         os.environ["HTTP_RESPONSE_HEADERS"] = json.dumps(
             {"Referrer-Policy": "no-referrer-when-downgrade"}
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertIn(
             "add_header Referrer-Policy 'no-referrer-when-downgrade';",
             header_config,
@@ -43,14 +43,14 @@ class TestCaseCustomHeaderConfig(unittest.TestCase):
         os.environ["HTTP_RESPONSE_HEADERS"] = json.dumps(
             {"Referrer-Policy": "no-referrr-when-downgrade"}
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertEquals("", header_config)
 
     def test_valid_header_accessControl(self):
         os.environ["HTTP_RESPONSE_HEADERS"] = json.dumps(
             {"Access-Control-Allow-Origin": "*"}
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertIn(
             "add_header Access-Control-Allow-Origin '*';", header_config
         )
@@ -59,14 +59,14 @@ class TestCaseCustomHeaderConfig(unittest.TestCase):
         os.environ["HTTP_RESPONSE_HEADERS"] = json.dumps(
             {"Access-Control-Allow-Origin": "htps://this.is.mydomain.nl"}
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertEquals("", header_config)
 
     def test_valid_header_contentType(self):
         os.environ["HTTP_RESPONSE_HEADERS"] = json.dumps(
             {"X-Content-Type-Options": "nosniff"}
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertIn(
             "add_header X-Content-Type-Options 'nosniff';", header_config
         )
@@ -75,7 +75,7 @@ class TestCaseCustomHeaderConfig(unittest.TestCase):
         os.environ["HTTP_RESPONSE_HEADERS"] = json.dumps(
             {"X-Content-Type-Options": ""}
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertEquals("", header_config)
 
     def test_valid_header_contentSecurity(self):
@@ -84,7 +84,7 @@ class TestCaseCustomHeaderConfig(unittest.TestCase):
                 "Content-Security-Policy": "default-src https: \u0027unsafe-eval\u0027 \u0027unsafe-inline\u0027; object-src \u0027none\u0027"  # noqa: E501
             }
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertIn(
             "add_header Content-Security-Policy 'default-src https: \\'unsafe-eval\\' \\'unsafe-inline\\'; object-src \\'none\\'';",  # noqa: E501
             header_config,
@@ -96,14 +96,14 @@ class TestCaseCustomHeaderConfig(unittest.TestCase):
                 "Content-Security-Policy": "$# default-src https://my.csp.domain.amsterdam"
             }
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertEquals("", header_config)
 
     def test_valid_header_permittedPolicies(self):
         os.environ["HTTP_RESPONSE_HEADERS"] = json.dumps(
             {"X-Permitted-Cross-Domain-Policies": "by-content-type"}
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertIn(
             "add_header X-Permitted-Cross-Domain-Policies 'by-content-type';",
             header_config,
@@ -113,7 +113,7 @@ class TestCaseCustomHeaderConfig(unittest.TestCase):
         os.environ["HTTP_RESPONSE_HEADERS"] = json.dumps(
             {"X-Permitted-Cross-Domain-Policies": "#%#^#^"}
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertEquals("", header_config)
 
     def test_valid_header_xssProtection(self):
@@ -122,7 +122,7 @@ class TestCaseCustomHeaderConfig(unittest.TestCase):
                 "X-XSS-Protection": "1; report=https://domainwithnewstyle.tld.consultancy"
             }
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertIn(
             "add_header X-XSS-Protection '1; report=https://domainwithnewstyle.tld.consultancy';",
             header_config,
@@ -132,7 +132,7 @@ class TestCaseCustomHeaderConfig(unittest.TestCase):
         os.environ["HTTP_RESPONSE_HEADERS"] = json.dumps(
             {"X-XSS-Protection": "1;mode=bock"}
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertEquals("", header_config)
 
     def test_valid_header_partial(self):
@@ -143,7 +143,7 @@ class TestCaseCustomHeaderConfig(unittest.TestCase):
                 "X-Content-Type-Options": "nosniff",
             }
         )
-        header_config = start.parse_headers()
+        header_config = headers.parse_headers()
         self.assertNotIn(
             "add_header X-XSS-Protection '1; report=https://domainwithnewstyle.tld.consultancy';",
             header_config,
@@ -152,4 +152,4 @@ class TestCaseCustomHeaderConfig(unittest.TestCase):
     def test_invalid_header_json(self):
         os.environ["HTTP_RESPONSE_HEADERS"] = "invalid"
         with self.assertRaises(ValueError):
-            start.parse_headers()
+            headers.parse_headers()
